refactor(apps/pano): rename folders
