+++ apps/kampus/app/pano/page.tsx
@@ -1,6 +1,6 @@
 import { redirect } from "next/navigation";
 
-import { DEFAULT_FILTER_PATH } from "~/app/pano/features/filter/utils";
+import { DEFAULT_FILTER_PATH } from "~/app/pano/features/PostFilter/utils";
 
 export default function PanoHome() {
   redirect(DEFAULT_FILTER_PATH);
